Refactor after CR
