--- scripts/release/const.ts
@@ -1,4 +1,0 @@
-export const base = '../../packages/';
-export const registry = 'https://registry.npmjs.org';
-// export const registry = 'http://localhost:4873';
-// npm publish --registry ${registry}
