+++ scripts/release/constants.ts
@@ -1,3 +1,4 @@
+import path from 'path';
 
 enum PACKAGE_TYPES {
     NotPackage = 0,
@@ -26,4 +27,9 @@ interface OperationsData {
     oldFiles: Record<string, any>;
 }
 
-export { PACKAGE_TYPES, PACKAGE_SUBTYPE, RELEASE_GRADATIONS, OperationsData };
+const BASE = path.resolve('../packages/');
+const REGISTRY = 'https://registry.npmjs.org';
+// const REGISTRY = 'http://localhost:4873';
+// npm publish --registry ${registry}
+
+export { BASE, REGISTRY, PACKAGE_TYPES, PACKAGE_SUBTYPE, RELEASE_GRADATIONS, OperationsData };
